backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0
pymongo==4.7.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
Pillow==10.3.0
pytesseract==0.3.10
easyocr==1.7.1
torch==2.3.0
torch-geometric==2.5.3
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
loguru==0.7.2
pytest==8.2.1
pytest-asyncio==0.23.7
boto3==1.34.110
firebase-admin==6.5.0
tenacity==8.3.0
aiofiles==23.2.1
python-slugify==8.0.4
